requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
python-dotenv>=0.19
Werkzeug>=2.0 # Usually comes with Flask
itsdangerous>=2.0
MarkupSafe>=2.0 # Usually comes with Flask
# Add other dependencies if you use them (e.g., Flask-Login, Flask-Migrate, Flask-Mail)